requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.4.2
openai==1.2.4
python-multipart==0.0.6
jinja2==3.1.2
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
